+++ fabric/src/main/java/com/jozufozu/flywheel/FlywheelFabric.java
@@ -1,0 +1,10 @@
+package com.jozufozu.flywheel;
+
+import net.fabricmc.api.ClientModInitializer;
+
+public class FlywheelFabric implements ClientModInitializer {
+	@Override
+	public void onInitializeClient() {
+
+	}
+}
